+++ NEWS.md
@@ -10,7 +10,7 @@
 - Fix caching errors in `MapDist()`
   ([#98](https://github.com/ms609/TreeDist/issues/98)).
 
-- Update tests for compatability with ape 5.7
+- Update tests for compatibility with ape 5.7
 
 
 # TreeDist 2.5.0